+++ CHANGELOG.md
@@ -2,6 +2,8 @@ CHANGELOG for Sulu
 ==================
 
 * dev-develop
+    * BUGFIX      #3206 [SnippetBundle]       Corrected translations in copy locale and open ghost overlay
+    * ENHANCEMENT #3206 [ContentBundle]       Made translations for copy locale and open ghost overlay changeable
     * BUGFIX      #3200 [SecurityBundle]      Fixed broken UserManager when used without Security
     * BUGFIX      #3183 [ContentBundle]       Fixed Windows xinclude error